Move sound templatetag tests to separate file
https://github.com/MTG/freesound/pull/1364
